+++ src/main/java/br/com/zalf/prolog/webservice/colaborador/ColaboradorService.java
@@ -186,6 +186,7 @@ public class ColaboradorService {
                 final ControleIntervaloDao dao = Injection.provideControleIntervaloDao();
                 final List<TipoIntervalo> tiposIntervalo = dao.getTiposIntervalosByUnidade(
                         colaborador.getUnidade().getCodigo(),
+                        Boolean.TRUE,
                         true);
                 loginHolder.setTiposIntervalos(tiposIntervalo);
             }