+++ src/main/java/br/com/zalf/prolog/webservice/gente/controleintervalo/ControleIntervaloDao.java
@@ -42,6 +42,7 @@ public interface ControleIntervaloDao {
 
     @NotNull
     List<TipoIntervalo> getTiposIntervalosByUnidade(@NotNull final Long codUnidade,
+                                                    final Boolean apenasAtivos,
                                                     final boolean withCargos) throws SQLException;
 
     @NotNull