+++ src/main/java/br/com/zalf/prolog/webservice/gente/controleintervalo/DeprecatedControleIntervaloResource.java
@@ -104,7 +104,7 @@ public class DeprecatedControleIntervaloResource {
             Pilares.Gente.Relatorios.INTERVALOS})
     @Path("/tipos/{codUnidade}/completos")
     public List<TipoIntervalo> getTiposIntervalosCompletos(@PathParam("codUnidade") Long codUnidade) {
-        return service.getTiposIntervalos(codUnidade, true);
+        return service.getTiposIntervalos(codUnidade, Boolean.TRUE,true);
     }
 
     @POST
@@ -141,7 +141,7 @@ public class DeprecatedControleIntervaloResource {
             Pilares.Gente.Intervalo.CRIAR_TIPO_INTERVALO})
     @Path("/tipos/{codUnidade}/resumidos")
     public List<TipoIntervalo> getTiposIntervalosResumidos(@PathParam("codUnidade") Long codUnidade) {
-        return service.getTiposIntervalos(codUnidade, false);
+        return service.getTiposIntervalos(codUnidade, Boolean.TRUE,false);
     }
 
     /**