+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Change Log
 
 ### Features
 * Cria DummyResource para retornar o JSON padrão do objeto Cargo
+* Permite filtrar por tipos de intervalo ativos/inativos ao realizar a buscas dos tipos de uma unidade
 
 ### Refactors
 * Altera método de inativação de tipo de intervalo para permitir ativar/inativar